added default syntax inference from extension. Mechanism to build namespace for Python syntax checking is introduced. It made extensible so that Petra-M can inject its own namespace for syntax checking
